chore(*): make IBaseSearchInfo internal
